* #62: Refactored response consumers.
